requirements.txt
attrs==22.2.0
brotli==1.0.9
dj-rest-auth==3.0.0
django-allauth==0.53.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==22.1
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
django-rest-auth==0.9.5
djoser==2.1.0
drf-yasg==1.21.5
google-api-core==2.11.0
google-auth-httplib2==0.1.0
mutagen==1.46.0
pdf2image==1.16.3
phonenumbers==8.13.7
pip-chill==1.0.1
pisa==3.0.33
psycopg2==2.9.5
python-dotenv==1.0.0
weasyprint==58.1
xhtml2pdf==0.2.9
zopfli==0.2.2
